// src/hoc/withField.d.ts
interface WithFieldProps {
  name: string;
  id?: string;
  formGroupClass?: string;
  required?: boolean;
  optionalLabel?: string;
  error?: string;
  description?: string;
  afterContent?: any;
  capitalize?: boolean;
  formGroupLabel?: string;
  label?: string;
  withFormGroup?: boolean;
  _formGroupProps?: object;
}